Refactor: Purge legacy HVAC commands and reinstate exhaustive edge-case builder tests - Purge src/ramses_tx/command/hvac.py and HvacMixins - Remove HVAC routing from CODE_API_MAP - Delete legacy test_apis_hvac.py - Reinstate exhaustive 1298, 12A0, 22F1, 22F7, and 31DA parameterized test payloads into test_hvac_builders.py using CQRS Intents - Clean up legacy tests in test_command.py - Fix broken Command.get_fan_param mock in test_hvac_ventilator.py - Consolidate parity tests into test_tx_parity.py
