+++ src/ramses_tx/command/__init__.py
@@ -6,12 +6,11 @@ from __future__ import annotations
 from ..const import I_, RP, RQ, W_, Code
 from .base import CommandBase
 from .dhw import DhwMixins
-from .hvac import HvacMixins
 from .system import SystemMixins
 from .zones import ZoneMixins
 
 
-class Command(DhwMixins, HvacMixins, SystemMixins, ZoneMixins, CommandBase):
+class Command(DhwMixins, SystemMixins, ZoneMixins, CommandBase):
     """The Command class (packets to be transmitted).
 
     They have QoS and/or callbacks (but no RSSI).
@@ -24,18 +23,12 @@ CODE_API_MAP = {
     f"{I_}|{Code._3EF0}": Command.put_actuator_state,
     f"{I_}|{Code._1FC9}": Command.put_bind,
     f"{W_}|{Code._1FC9}": Command.put_bind,
-    f"{W_}|{Code._22F7}": Command.set_bypass_position,
-    f"{I_}|{Code._1298}": Command.put_co2_level,
     f"{RQ}|{Code._1F41}": Command.get_dhw_mode,
     f"{W_}|{Code._1F41}": Command.set_dhw_mode,
     f"{RQ}|{Code._10A0}": Command.get_dhw_params,
     f"{W_}|{Code._10A0}": Command.set_dhw_params,
     f"{RQ}|{Code._1260}": Command.get_dhw_temp,
     f"{I_}|{Code._1260}": Command.put_dhw_temp,
-    f"{I_}|{Code._22F1}": Command.set_fan_mode,
-    f"{W_}|{Code._2411}": Command.set_fan_param,
-    f"{RQ}|{Code._2411}": Command.get_fan_param,
-    f"{I_}|{Code._12A0}": Command.put_indoor_humidity,
     f"{RQ}|{Code._1030}": Command.get_mix_valve_params,
     f"{W_}|{Code._1030}": Command.set_mix_valve_params,
     f"{RQ}|{Code._3220}": Command.get_opentherm_data,
@@ -65,5 +58,4 @@ CODE_API_MAP = {
     f"{W_}|{Code._2309}": Command.set_zone_setpoint,
     f"{RQ}|{Code._30C9}": Command.get_zone_temp,
     f"{RQ}|{Code._12B0}": Command.get_zone_window_state,
-    f"{I_}|{Code._31DA}": Command.get_hvac_fan_31da,
 }
