+++ tests/tests_rf/commands/test_tx_parity.py
@@ -248,169 +248,3 @@ def test_build_get_opentherm_data() -> None:
     assert dto.addr1 == legacy_cmd._addrs[0].id
     assert dto.addr2 == legacy_cmd._addrs[1].id
     assert dto.addr3 == legacy_cmd._addrs[2].id
-
-
-def test_build_put_co2_level() -> None:
-    legacy_cmd = LegacyCommand.put_co2_level("32:111111", 400.0)
-    intent = Command(
-        src=Address("32:111111"),
-        dst=Address("32:111111"),
-        action=Action.PUT_CO2_LEVEL,
-        data={"co2_level": 400.0},
-    )
-    dto = build_dto(intent)
-    assert dto.verb == legacy_cmd.verb
-    assert dto.code == legacy_cmd.code
-    assert dto.payload == legacy_cmd.payload
-    assert dto.addr1 == legacy_cmd._addrs[0].id
-    assert dto.addr2 == legacy_cmd._addrs[1].id
-    assert dto.addr3 == legacy_cmd._addrs[2].id
-
-
-def test_build_put_indoor_humidity() -> None:
-    legacy_cmd = LegacyCommand.put_indoor_humidity("32:111111", 0.5)
-    intent = Command(
-        src=Address("32:111111"),
-        dst=Address("32:111111"),
-        action=Action.PUT_INDOOR_HUMIDITY,
-        data={"indoor_humidity": 0.5},
-    )
-    dto = build_dto(intent)
-    assert dto.verb == legacy_cmd.verb
-    assert dto.code == legacy_cmd.code
-    assert dto.payload == legacy_cmd.payload
-    assert dto.addr1 == legacy_cmd._addrs[0].id
-    assert dto.addr2 == legacy_cmd._addrs[1].id
-    assert dto.addr3 == legacy_cmd._addrs[2].id
-
-
-def test_build_set_fan_mode() -> None:
-    legacy_cmd = LegacyCommand.set_fan_mode(
-        "32:111111", fan_mode="low", scheme="itho", src_id="18:000730"
-    )
-    intent = Command(
-        src=Address("18:000730"),
-        dst=Address("32:111111"),
-        action=Action.SET_FAN_MODE,
-        data={"fan_mode": "low", "scheme": "itho"},
-    )
-    dto = build_dto(intent)
-    assert dto.verb == legacy_cmd.verb
-    assert dto.code == legacy_cmd.code
-    assert dto.payload == legacy_cmd.payload
-    assert dto.addr1 == legacy_cmd._addrs[0].id
-    assert dto.addr2 == legacy_cmd._addrs[1].id
-    assert dto.addr3 == legacy_cmd._addrs[2].id
-
-
-def test_build_set_bypass_position() -> None:
-    legacy_cmd = LegacyCommand.set_bypass_position(
-        "32:111111", bypass_mode="auto", src_id="18:000730"
-    )
-    intent = Command(
-        src=Address("18:000730"),
-        dst=Address("32:111111"),
-        action=Action.SET_BYPASS_POSITION,
-        data={"bypass_mode": "auto"},
-    )
-    dto = build_dto(intent)
-    assert dto.verb == legacy_cmd.verb
-    assert dto.code == legacy_cmd.code
-    assert dto.payload == legacy_cmd.payload
-    assert dto.addr1 == legacy_cmd._addrs[0].id
-    assert dto.addr2 == legacy_cmd._addrs[1].id
-    assert dto.addr3 == legacy_cmd._addrs[2].id
-
-
-def test_build_get_fan_param() -> None:
-    legacy_cmd = LegacyCommand.get_fan_param("32:111111", "31", src_id="18:000730")
-    intent = Command(
-        src=Address("18:000730"),
-        dst=Address("32:111111"),
-        action=Action.GET_FAN_PARAM,
-        data={"param_id": "31"},
-    )
-    dto = build_dto(intent)
-    assert dto.verb == legacy_cmd.verb
-    assert dto.code == legacy_cmd.code
-    assert dto.payload == legacy_cmd.payload
-    assert dto.addr1 == legacy_cmd._addrs[0].id
-    assert dto.addr2 == legacy_cmd._addrs[1].id
-    assert dto.addr3 == legacy_cmd._addrs[2].id
-
-
-def test_build_set_fan_param() -> None:
-    legacy_cmd = LegacyCommand.set_fan_param("32:111111", "31", 30, src_id="18:000730")
-    intent = Command(
-        src=Address("18:000730"),
-        dst=Address("32:111111"),
-        action=Action.SET_FAN_PARAM,
-        data={"param_id": "31", "value": 30},
-    )
-    dto = build_dto(intent)
-    assert dto.verb == legacy_cmd.verb
-    assert dto.code == legacy_cmd.code
-    assert dto.payload == legacy_cmd.payload
-    assert dto.addr1 == legacy_cmd._addrs[0].id
-    assert dto.addr2 == legacy_cmd._addrs[1].id
-    assert dto.addr3 == legacy_cmd._addrs[2].id
-
-
-def test_build_get_hvac_fan_31da() -> None:
-    legacy_cmd = LegacyCommand.get_hvac_fan_31da(
-        "32:111111",
-        "0000",
-        bypass_position=None,
-        air_quality=None,
-        co2_level=None,
-        indoor_humidity=None,
-        outdoor_humidity=None,
-        exhaust_temp=None,
-        supply_temp=None,
-        indoor_temp=None,
-        outdoor_temp=None,
-        speed_capabilities=None,
-        fan_info=None,
-        _unknown_fan_info_flags=[],
-        exhaust_fan_speed=None,
-        supply_fan_speed=None,
-        remaining_mins=None,
-        post_heat=None,
-        pre_heat=None,
-        supply_flow=None,
-        exhaust_flow=None,
-    )
-    intent = Command(
-        src=Address("32:111111"),
-        dst=Address("32:111111"),
-        action=Action.GET_HVAC_FAN_31DA,
-        data={
-            "hvac_id": "0000",
-            "bypass_position": None,
-            "air_quality": None,
-            "co2_level": None,
-            "indoor_humidity": None,
-            "outdoor_humidity": None,
-            "exhaust_temp": None,
-            "supply_temp": None,
-            "indoor_temp": None,
-            "outdoor_temp": None,
-            "speed_capabilities": None,
-            "fan_info": None,
-            "_unknown_fan_info_flags": [],
-            "exhaust_fan_speed": None,
-            "supply_fan_speed": None,
-            "remaining_mins": None,
-            "post_heat": None,
-            "pre_heat": None,
-            "supply_flow": None,
-            "exhaust_flow": None,
-        },
-    )
-    dto = build_dto(intent)
-    assert dto.verb == legacy_cmd.verb
-    assert dto.code == legacy_cmd.code
-    assert dto.payload == legacy_cmd.payload
-    assert dto.addr1 == legacy_cmd._addrs[0].id
-    assert dto.addr2 == legacy_cmd._addrs[1].id
-    assert dto.addr3 == legacy_cmd._addrs[2].id
