+++ tests/tests_tx/test_command.py
@@ -255,144 +255,3 @@ async def test_rq_missing_target() -> None:
             addr0="--:------",
             addr1="--:------",
         )
-
-
-# --- set_fan_mode (22F1) scheme-aware payload tests (issue #547) ---
-
-
-def test_set_fan_mode_orcon_2byte_default() -> None:
-    """Default scheme (orcon) produces a 3-byte payload with mode_max=07."""
-    # Arrange
-    expected_payload = "000107"
-    expected_code = "22F1"
-
-    # Act
-    cmd = Command.set_fan_mode("37:111111", "low", src_id="18:000730")
-
-    # Assert
-    # ORCON: low=01, mode_max=07 -> payload "000107"
-    assert cmd.payload == expected_payload
-    assert str(cmd.code) == expected_code
-
-
-def test_set_fan_mode_orcon_2byte_explicit() -> None:
-    """Explicit orcon scheme with mode_max='' produces legacy 2-byte payload."""
-    # Arrange
-    expected_payload = "0001"
-
-    # Act
-    cmd = Command.set_fan_mode(
-        "37:111111",
-        "low",
-        scheme="orcon",
-        src_id="18:000730",
-        legacy_format=True,
-    )
-
-    # Assert
-    # 2-byte legacy form: idx + mode only
-    assert cmd.payload == expected_payload
-
-
-def test_set_fan_mode_itho_3byte() -> None:
-    """Itho scheme produces 3-byte payload with mode_max=04."""
-    # Arrange
-    expected_payload = "000204"
-
-    # Act
-    cmd = Command.set_fan_mode("37:111111", "low", scheme="itho", src_id="18:000730")
-
-    # Assert
-    # ITHO: low=02, mode_max=04 -> "000204"
-    assert cmd.payload == expected_payload
-
-
-def test_set_fan_mode_vasco_3byte() -> None:
-    """Vasco scheme produces 3-byte payload with mode_max=06."""
-    # Arrange
-    expected_payload = "000406"
-
-    # Act
-    cmd = Command.set_fan_mode("37:111111", "high", scheme="vasco", src_id="18:000730")
-
-    # Assert
-    # VASCO: high=04, mode_max=06 -> "000406"
-    assert cmd.payload == expected_payload
-
-
-def test_set_fan_mode_siber_3byte_from_issue() -> None:
-    """Siber DF Evo 4 payloads from issue #547 (orcon scheme, mode_max=07)."""
-    # Arrange
-    expected_low = "000107"
-    expected_int = "000207"
-
-    # Act
-    # low:  003 000207
-    cmd_low = Command.set_fan_mode(
-        "37:111111", "low", scheme="orcon", src_id="18:000730"
-    )
-
-    # The issue lists Siber modes as 02=low,03=medium,04=high,07=boost.
-    # With orcon scheme these map to: 02=medium,03=high,04=auto,07=off.
-    # For Siber, the user should use the itho scheme or raw int indices.
-    # Verify int index works: 02 -> "000207"
-    cmd_int = Command.set_fan_mode(
-        "37:111111", 0x02, scheme="orcon", src_id="18:000730"
-    )
-
-    # Assert
-    assert cmd_low.payload == expected_low
-    assert cmd_int.payload == expected_int
-
-
-def test_set_fan_mode_nuaire_3byte() -> None:
-    """Nuaire scheme produces 3-byte payload with mode_max=0A."""
-    # Arrange
-    expected_payload = "00020A"
-
-    # Act
-    cmd = Command.set_fan_mode(
-        "37:111111", "normal", scheme="nuaire", src_id="18:000730"
-    )
-
-    # Assert
-    # NUAIRE: normal=02, mode_max=0A -> "00020A"
-    assert cmd.payload == expected_payload
-
-
-def test_set_fan_mode_int_index() -> None:
-    """Integer fan_mode is treated as a hex mode index."""
-    # Arrange
-    expected_payload = "000307"
-
-    # Act
-    cmd = Command.set_fan_mode("37:111111", 3, scheme="orcon", src_id="18:000730")
-
-    # Assert
-    assert cmd.payload == expected_payload
-
-
-def test_set_fan_mode_none_is_off() -> None:
-    """None fan_mode maps to mode 00 (off/away)."""
-    # Arrange
-    expected_payload = "000007"
-
-    # Act
-    cmd = Command.set_fan_mode("37:111111", None, scheme="orcon", src_id="18:000730")
-
-    # Assert
-    assert cmd.payload == expected_payload
-
-
-def test_set_fan_mode_invalid_scheme_raises() -> None:
-    """An unknown scheme raises CommandInvalid."""
-    # Arrange & Act & Assert
-    with pytest.raises(CommandInvalid, match="scheme is not valid"):
-        Command.set_fan_mode("37:111111", "low", scheme="bogus", src_id="18:000730")
-
-
-def test_set_fan_mode_invalid_mode_raises() -> None:
-    """A mode not in the scheme's map raises CommandInvalid."""
-    # Arrange & Act & Assert
-    with pytest.raises(CommandInvalid, match="fan_mode is not valid"):
-        Command.set_fan_mode("37:111111", "turbo", scheme="itho", src_id="18:000730")
